fix: resolve C++ compilation errors in unit tests
- Fix size_t to int conversion warnings in test_main.cxx/hxx
- Simplify reward test to avoid mock class API mismatches
- Remove problematic Aircraft/Missile mock constructors

+++ bvr_sim/src_cxx/test_main.hxx
@@ -94,7 +94,7 @@ private:
     size_t current_test_idx = 0;
 
     void print_summary() {
-        int passed = 0, total = tests.size();
+        size_t passed = 0, total = tests.size();
         for (const auto& test : tests) {
             if (test.passed) passed++;
         }
